requirements.txt
python-dotenv
JSONSchema2DB==1.0.1
psycopg2
numpy
opencv-contrib-python
imutils
scikit-image
# psycopg2==2.7.2
